Remove `base_path` from `world_location` links in news articles
We were including the `base_path` attribute for `world_location` links
in news articles, which was not included in the Content Store version of
the page.

The frontend code [1] forms a URL based on either the `base_path` (if
available) or the link text.

As the `base_path` was included, we were linking to a different URL on
the two versions of news articles.

Removing the `base_path` so we fallback to using the title, in the same
way as this is done for the Content Store version of the page.

1: https://github.com/alphagov/govuk_publishing_components/blob/ea2125f3b274e8d50b51f957ddbc0dcc4c4c61fb/lib/govuk_publishing_components/presenters/related_navigation_helper.rb#L126C1-L135C10