Only send redirects if the previous base path is present
I think this is the first time we’ve introduced a base_path to an
existing schema type, and, as a result, there may be circumstances
where a block exists without a path. If this is the case, we don’t
want to create a redirect, because there would be nothing to redirect
to.

To get around this, we first check if the `previous_base_path` is not
`nil` or `empty` before checking the `previous_base_path` and
`base_path` are the same