Hand-write SQL for dataloaders
This replaces the ActiveRecord generated SQL in the previous commit with
hand written SQL.

By passing the parameters through as a JSON string and calling
jsonb_to_recordset on it, we avoid the need to call .sub on the SQL
string.

The intent in this commit is to retain the existing behaviour, so no
tests have been changed.