Add benchmarks using RSpec
We used to have a set of benchmarks for various important queries, until
they were removed in 48c36e3e (because we hadn't needed them in a while,
and they had linting issues).

This commit reintroduces benchmarks, but wraps them up using RSpec
instead of custom scripts.

It's important we don't clash with the existing specs, because they
clean the database on every run and the benchmarks need a realistic
database (i.e. a copy of integration's database) and it's very annoying
if that gets accidentally cleaned (as it takes so long to download). To
avoid clashing, I've used a pattern of `name_benchmark.rb` instead of
`name_spec.rb` for the benchmarks. These then use a different
.rspec.benchmark set of options, and have a separate benchmark_helper.rb
spec helper (which doesn't clean the database). I've introduced a new
database connection environment (:benchmark), which is the same as
development but uses a readonly connection so we don't accidentally
trash the data - these benchmarks should only do selects, not inserts /
updates.

The benchmarks should be run with:

    bundle exec rspec --options .rspec.benchmark

I've introduced some helpers and matchers to make it easy to write tests
which:
- Check which SQL queries have been made and how long they took
- Run EXPLAIN ANALYZE on any slow queries and print the plans in the
  failure message
- Profile the code with StackProf, and if it's too slow write the
  profile into tmp/ so it can be analyzed with a flamegraph viewer (like
  https://speedscope.app)

I'm pretty happy with it.