Add boilerplate for RSpec benchmarks
These should run independently of the main specs, so they use a
different pattern (*_benchmark.rb instead of the default *_spec.rb).

They point at the benchmark database configuration (i.e. the same
database as development), so they don't attempt to clean the database
before the suite runs.

They create a few directories for reports which we're not using yet.
They also do a bit of ActiveRecord warm up so that the first test
doesn't hit a bunch of class loading / initialisation code which isn't
relevant to performance for a warm app.