Custom RSpec matcher to check a SQL query is fast
This is a bit fancy... if the SQL query is slower than the threshold,
the failure matcher will automatically run an EXPLAIN (ANALYZE, BUFFERS)
for the query, and store the query and the plan in files in
tmp/benchmarks.

I think this is useful, because it lowers the barrier to entry to
looking at SQL query plans. Of course, it doesn't make them easier to
read, but that's kinda unavoidable.