Add RSpec formatter to report benchmark stats
This will output results like:

    PublishingApiSchema
      should efficiently render ministers_index example - /government/ministers
        Overall time 1.152 | SQL 0.165, Govspeak 0, Other 0.987 | 7 SQL queries (max: 0.123, min: 0.0, avg: 0.024) | 0 Govspeak renders ()
          Timings report: tmp/benchmarks/instrumentation/should-efficiently-render-ministers_index-example-government-ministers.json
          CPU Profile: tmp/benchmarks/profiles/should-efficiently-render-ministers_index-example-government-ministers.json

Which makes it easy to see at a glance how a particular bit of code
performed (and often why it was slow - e.g. a slow bit of SQL, a large
bit of govspeak).

It's forked from RSpec's DocumentationFormatter, because that one's the
closest to having nice output to start with. As far as I can tell, the
DocumentationFormatter isn't meant to be extended - the "right thing to
do" is to extend BaseTextFormatter, so I've had to copy a fair bit of
implementation over from RSpec. I think it's just about worth it though.