Add benchmarks for some graphql example queries
I've put the test cases in a YAML file, as eventually we could have
quite a lot of these. The test code just loops over the examples in the
file, and checks that the queries for those base paths execute within
the thresholds given.

It's important that the thresholds are generous enough that the
benchmarks pass on modest hardware (like this ancient intel macbook I'm
still stuck with) - they'd run faster in a production like environment.