Remove redundant check
The result of this line of code was being ignored, so it can safely be
removed.

It seems like it was either committed by mistake or was missing an OR
between it and the follow line, offering an early exit for
straightforward path matches. In my local env benchmarking that early
exit didn't appear to make any significant performance improvement, so
it might as well go.