Update test example of unsupported content type
This tests the behaviour of the #live_content endpoint in cases where
the requested content's type isn't one that the endpoint supports.

We'll be adding support for travel_advice_index shortly, so I'm bumping
this again to an example of a content type that I know I won't be adding
support for (because there don't appear to be any live Editions with the
get_involved schema_name and I've used this fact to justify trimming it
off of my list).