Insert auto reverse links in GraphQL content item
During link expansion, LinkExpansion#links_with_content inserts auto
reverse links (or it performs something called auto reverse linking ?):
the result is that all top-level reverse links have a link added to them
that points back to the root edition via the corresponding inverse
link_type, e.g.

  Edition A -document_collections-> Edition B -documents-> Edition A

This update to the GraphqlContentItemService reproduces that behaviour
for content items served from the `/graphql/content` endpoint.

Note: I went down this route before fully considering the alternative
option of updating the query builder and GraphQL schema to support
querying a link_type for a specific linked edition. It'd probably be
more work than this was, but I wonder if it might be less messy in
a way.