Add a `type` to the BigQuery credentials
It seems that a recent change to the Google Cloud Ruby gems has meant
that our usual method of creating a keyfile from a hash (rather than
reading from a static JSON file, as Google insist on) wasn’t working,
because the `type` attribute was missing:

```
The provided credentials were not of type 'service_account'. Instead,
the type was ''
```

Adding a `type` attribute to our `credential` JSON “file” seems to fix
the issue, and I’ve confirmed this in the console in production.