Correctly check editions ordered by position
Previously this test set the order up in the same way as the link ids,
so the editions would have been ordered that way even if they all had
the same position.

Worse, we were using match_array, which doesn't actually check the
ordering :facepalm:, so the test would have passed whatever the ordering
was.