Use eq instead of match_array where order is clear
We can use match_array when we just want to check the actual array
contains the same items as the expected, but we don't care about the
order. If we have fewer than two items, or if the test is checking a
specific order, it's better to be stricter and use eq.