Rename and sort benchmark examples
Rename = update the comments to more closely reflect the schema_name.
Sort = alphabetically. Maybe ministers index was at the top for a good
reason, though, since it's an interesting example. Anyhow, right now I'm
en route to adding a lot more examples to this file and this makes it
clearer how they should be named and where they should go.