Parameterize query inputs using json_to_recordset
ActiveRecord doesn't provide a way to pass an array of arrays as a bind
parameter, so we were previously doing a string replacement on the query
(which feels very much like bad practice).

ActiveRecord can pass strings as bind parameters though, and Postgres
can parse JSON. This commit passes a JSON string to the query, and then
turns it into SQL types using json_to_recordset:

    https://www.postgresql.org/docs/current/functions-json.html#id-1.5.8.22.8.13.2.2.13.1.1.1

This actually simplifies the work the query planner has to do.
Previously it worked out that it could build do one index scan per
(edition_id, link_type) tuple, then bitmap-OR them all together. That
was fast to execute, but slow to plan for large numbers of tuples.

Now the planner just knows it's going to get a CTE with id / link_type
columns, and that it needs to use these in a JOIN. It still manages to
do this with a series of nested loop index scans (which are fast to
execute), but the planning time is significantly reduced because it's
not thinking about large lists of hardcoded tuples anymore.

Note that I've also passed query_input_count as a parameter, and used
this to LIMIT the query_input CTE. This is because the postgresql
planner doesn't know how many rows json_to_recordset will return (as it
depends on the input), and therefore always assumes it will return 100
rows. In practice, we usually have far fewer inputs than this. Adding a
LIMIT to the CTE lets the planner know how many rows it can really
expect, which should reduce the risk of bad plans based on inaccurate
assumptions.