Remove redundant field from jsonnet
The `withdrawn_notice` field seems to have been included in these
schemas by accident, under the details field. It has the same name and
definition as a default top-level field. No existing editions use it.

If it were to stay in the jsonnet schemas, then it would have to be
added to app/graphql/types/edition_type.rb for graphql queries of those
schemas to work.