Ensure forbidden details field exists
GraphQL doesn’t support empty objects `{}` but the special_route jsonnet
schema forces that to be the value of the `details` field, and the
frontend schemas derived from jsonnet require the field to exist.

* `special_route` jsonnet specifies details: "forbidden". For some
  reason, `GovukSchemas::Schema.find(frontend_schema: "special_route")`
  still thinks that the `details` field is required as
  `"details" => {"type" => "object", "additionalProperties" => false, "properties" => {}}`
  https://github.com/alphagov/publishing-api/blob/5173e32cb1e3dc446dcb45f30e00ae6119d47ebf/content_schemas/formats/special_route.jsonnet#L4)
* `/lib/schema_generator/format.rb` defines `"forbidden"` to be `{}`
  https://github.com/alphagov/publishing-api/blob/5173e32cb1e3dc446dcb45f30e00ae6119d47ebf/content_schemas/formats/special_route.jsonnet#L4
* GraphQL spec: https://spec.graphql.org/October2021/#sel-FAHZhCFDBAACDA4qe
* Ruby implementation: https://github.com/rmosolgo/graphql-ruby/pull/5137

This commit patches the ContentItemCompactor to create an empty
`details` field when it is 'required'. This occurs after graphql has
returned the content item, but before the tests that check that the
output schema has been matched.