Improve efficiency of withdrawn notices
Calling Presenters::EditionPresenter is _extremely inefficient_ - as
well as getting the withdrawn_notice (if there is one) it does full link
expansion on the edition, using the unoptimised link expansion code we
use at publish time (when presenting an edition to content store).

The only place we're using the presented edition is to generate this
hash with two things in - explanation and withdrawn_at. The code isn't
entirely trivial, but it's simple enough that I don't feel too bad
duplicating it.

This reduces the number of SQL queries for /government/news/public-exhibitions-for-a21-safety-improvements-in-east-sussex
from 235 to 11.