Don't mistake Sentry::ErrorEvent for content reference
We were accidentally returning the result of log_error from
replace_alias_content_id_with_content_id, which meant we ended up with
the Sentry::ErrorEvent in the middle of a list of content references.
This later failed with a NoMethodError when we tried to do
content_references.map(&:document_type).

https://govuk.sentry.io/issues/7036059169/events/139445a3ed0a4f85905e7fe04e0a4210

Always love to work with dynamically typed languages.

"You mistook the stars reflected on a pond at night for the sky"- Vilgerfortz of Roggeveen