Render govspeak in details when Putting content
As per [ADR-011], instead of rendering govspeak when presenting content
to content store, we're going to render it before storing the edition.
This means that instead of writing:

    details = [{content_type: "text/govspeak", content: "some content"}]

We'll write:

    details = [{content_type: "text/govspeak", content: "some content"}, {content_type: "text/html", content: "some content"}]

This will increase the size of the editions table, but we'll save some
rendering work when presenting editions (and when serving editions from
GraphQL).

[ADR-011]: https://github.com/alphagov/publishing-api/blob/main/docs/arch/adr-011-cache-rendered-govspeak.md