Add rendered_by and govspeak_version to rendered Govspeak
We want to pre-render Govspeak as part of a performance improvement for
GraphQL. In order to minimise the impact this has on the size of
Publishing API's database, we plan to remove the Govspeak we render in
Publishing API when superseding an edition. We therefore need to know
whether HTML content has been rendered by Publishing API (vs a
publishing app)

We haven't got a specific use in mind for govspeak_version, but it's something that
feels like it might be useful in the future (e.g. for re-rendering
everything on a bad govspeak version), and it's a small bit of data to
store.

Note that the regex in the test matches the major / minor / patch prefix
of a Semver version, but then allows any suffix with .* so it won't fail
if we use a version of govspeak with pre-release or build information in
the version number.