Refactor govspeak_details_renderer
We'd like to use this same class for another operation on content arrays
(removing content rendered by publishing-api). Before we do that,
refactoring it to be a bit more general, fixing some badly named methods
(looking at you parsed_content), and inlining a lot of overly small
methods.

The visit_content_arrays method takes a block, which it will call for
anything that looks like a "content array" (i.e. an array of {
content_type: ... } hashes). It's called "visit..." in a nod to the
Visitor Pattern - we're a little unsure whether this name might be
confusing, but we couldn't think of a better one.