Remove rendered govspeak when superseding editions
This is purely to reduce the amount of disk space we're taking up in the
editions table with rendered content. Given we'll still have the
original source govspeak, we're not losing any information by not
storing the rendered HTML for all time.

We're only removing things rendered_by publishing-api, rather than any
text/html paired with text/govspeak, because publishing-api couldn't
reliably reconstruct text/html that was rendered by a publishing app
with an unknown set of govspeak extensions (so that would lose
information).