Use pattern matching instead of if/else
I think this is a bit clearer.

It does change the semantics a little. Previously we were checking that
all the elements in an array had content_type on them, now we're just
checking the first. But I think that will be harmless in practice - we
don't expect to get arrays with some { content_type: ... } hashes and
some other things.