Add govspeak_version to rendered details content
We haven't got a specific use in mind for this, but it's something that
feels like it might be useful in the future (e.g. for re-rendering
everything on a bad govspeak version), and it's a small bit of data to
store.

Note that the regex in the test matches the major / minor / patch prefix
of a Semver version, but then allows any suffix with .* so it won't fail
if we use a version of govspeak with pre-release or build information in
the version number.