Fix route validation when paths contain regex special characters
It's legal for base_paths to include regex special characters like $

For example in /employment-tribunal-decisions/ms-g-adeji-v-miss-ca$h-sportswear-london-3200533-slash-2023

Concatenating these paths into regular expressions was causing incorrect
behaviour.

As far as I can understand the regular expressions, they look like they
were just a convoluted way of doing a "start_with?" so let's just
simplify the method to use that.

I've updated the test to use /ca$h-money as it's base path instead of
/vat-rates so that it covers this edge case. Just in case someone tries
to refuctor it in future.