Reduce duplicates found by embedded_content_finder
Previously if we had content like:

    [
      { content_type: "text/govspeak", content: "{{embed:...}}" },
      { content_type: "text/html", content: "{{embed:...}}" },
    ]

The embedded_content_finder would find two matches - one for the
text/govspeak content, one for the text/html content. As both of
these content hashes are supposed to be representations of the same
content, it makes sense only to look through one of them for matches.

To achieve this, I've special cased arrays containing at least one

    { content_type: "text/govspeak", ... }

... If we find one of those, we only search that one content hash, and
ignore all the others in the array.

This halves the number of duplicates we find in a couple of tests. One
test is described as

    "returns duplicates when there is more than one content reference in the field and #{field_name} is a multipart document"

But the name is still correct even with this change, as it creates
duplicates within each bit of content.