More efficient content id alias query
Currently, the CONTENT_ID_ALIAS_SQL subquery is inefficient.

    FROM content_id_aliases
    LEFT JOIN documents ON documents.id = subquery.document_id
    WHERE content_id_aliases.content_id = documents.content_id

Postgres decides the best way to do this is:

    for every document_id:
      scan the whole content_id_aliases table
      to check if the content_id exists

If instead, we ensure that content_id is selected in the original
subquery, we can do:

    FROM content_id_aliases
    WHERE content_id_aliases.content_id = subquery.content_id

And postgres only needs to scan the content_id_aliases table once, or
(if the table grows large enough for it to be worth it) it can use the
index on content_id.