Include content_id if state_history is requested
The LATERAL query to get state_history requires content_id to have been
selected.

This prevents an error if the user queries the state_history field but
doesn't include content_id.

I've also removed the fields variable, as I found it quite confusing to
have both an attr_reader called fields and a variable called fields used
in the same method. It's defined, used on the next line, and then
discareded, so feels fine to inline it.

I've also switched to a brace style block for a one liner - just think
it's nicer.