Ensure that details has symbol keys
There's a subtle difference with edition models between:

    edition[:details]

and

    edition.details

... the latter is a method defined by the SymbolizeJSON model, which
symbolizes any columns defined as JSONB. The former simply returns the
hash as it finds it (i.e. with string keys).

Previously we had to worry about situations where details might have
string keys or symbol keys - it depended on the code path.

Using the method instead of the hash lookup means we can simplify the
code by assuming symbols everywhere. In particular, this means we can
use pattern matching, which only works on symbol keys.