Make ContentTypeResolver more strict
At the moment, if you ask it for text/html, but give it an array that
_doesn't_ include text/html, you'll get the whole array back.

That's inevitably going to cause an error / problem down the line, so
it's much better to raise an error if it looks like we can't resolve the
content type we've been asked for.

This commit adds an error if it tries to extract content from an array
that looks like a content hash array, but doesn't include the
content_type we're looking for.