Skip embed change notes when edition has no embeds
Previously, when looking for change notes we would check all of the
embeds that had _ever_ been attached to a document, even if the current
version of the document didn't have any embeds.

Even though the query is quite fast, now that we've got the right
indexes in places, it's still expensive to look through the whole
document history when the vast majority of documents haven't got any
embeds at all.

This commit introduces a much cheaper check whether the _current_
edition has any embeds. If it doesn't, we don't try and fetch any
history from embeds that used to be attached to the document. If it
does, then we run the same query as before.

There are a couple of edge cases where I don't think this query behaves
in the way you'd probably want it to - for example, if a draft edition
of the document had an embed, but was never published, the history for
the embed would appear in the live edition's change history (I think).
And various cases around embeds that used to be on the document, but
aren't anymore. For now, I think those edge cases are just theoretical,
as usually once a document picks up an embed it tends to keep it.

I considered trying to filter just for the target editions that are
currently embedded, but this further complicates the query and I don't
think it's worth it unless we're worried about the issues in the
previous paragraph.