WIP Don't look for edition links in secondary locale
TODO: add relevant tests

In classic link expansion (the type used when preparing content items
for Content Store), we only look for links of a 'fallback' locale for
link set links. A fallback URL being for example 'en' when the
primary/root edition locale is 'cy'

In GraphQL we were looking for fallback locale links for both link set
and edition links. We want pages rendered using GraphQL-supplied content
items to look the same as those rendered using Content Store-supplied
JSON, at least in the short term, so this removes the fallback locale
for edition links

Note that we still include the `is_primary_locale` field - the `UNION
ALL` requires an equal number of columns between the UNIONed datasets