Allow HTML character entity quote in GraphQL diffs
I'm seeing &quot; in lieu of " in some frontend responses, which isn't
being captured and normalised by the datetime regex in our frontend
diffing functions. This makes it so that either quote representation is
allowed in the regex. Unfortunately we can't use non-capturing groups in
sed, so we just have to skip the unused group in the second half of the
expression