Rewrite GraphQL frontend diff script in Ruby
This is a simpler and more opinionated version of the Bash frontend
diffing script and its bulk wrapper. It makes used of the bulk check
wrapper that we use for bulk JSON diffs and other bulk processing, and
closely resembles the JSON diffing script

Differences from the Bash version:
- it always runs in development (custom host support will follow, which
  will allow you to run the diffs against other environments)
- the diff style is fixed (inline using git)
- it's non-interactive in all modes, not just bulk
- there's no option to skip the download in non-bulk mode
- in bulk mode it doesn't save diffs to disk - instead it follows the
  same interactive output approach as other users of the bulk check
  wrapper (stop when there's a diff, then continue when the user is
  ready)
- in bulk mode the input path is set

Ruby should be more familiar to the majority of our team than Bash, so
it'll hopefully be easier to extend and maintain

Since the HTML and JSON diffs are now have the same interface, we can
also restructure (hopefully simplify) the guidance. The integration
environment guidance is in an odd place in this commit, but the next
commit will fix that