Define SidekiqUniqueJobs lock_ttl as an integer
Apparently (https://github.com/mhenrixon/sidekiq-unique-jobs/issues/882,
https://github.com/mhenrixon/sidekiq-unique-jobs/pull/884#issuecomment-3154630380)
SidekiqUniqueJobs version 8.0.11 seems to have accidentally dropped
support for defining lock_ttl as an ActiveSupport::Duration. We do that.
And we updated to version 8.0.11 on 27th May this year
(https://github.com/alphagov/publishing-api/pull/3364).

We've been seeing an increase in problems with stale SidekiqUniqueJobs
locks these past few months and believed that we'd only just noticed the
problem recently because of digging deeper into the issue. I guess
there's a chance that this change is the source of those problems?

Either way, apprently it needs fixing so now we'll use an integer
representing seconds, which SidekiqUniqueJobs definitely supports.