Do not log a lack of draft edition to Sentry
In b729d3ee39bc1c7d00fe29e4a6b0ec99d8f5a062, we removed the logging of
`AbortWorkerError` from the `DownstreamLiveJob` to Sentry, since the
error is unactionable by us.

Doing the same here for the `DownstreamDraftJob`.

Since we do not use these logs for anything, this removes the logging to
Sentry and makes the tests more robust by checking that the content is
not pushed downstream.