Add details fields to landing_page graphql query
These fields are used by the handful of landing_page documents, and
affect the HTML that is eventually rendered.

They weren't included by the script that generated the graphql query
because they aren't mentioned in the schema, which is very loose. It
doesn't require any fields in `details`, and it allows any.