Add fields to world_location links in expansion rules
Depending on the schema of linked-to document,
govuk_publishing_components chooses or guesses a URL to render on the
website.

Targets of world_location links can be of the `world_location` schema,
or the `world_location_news` schema. If the target is
a `world_location`, then that document might itself have
a `world_location_news` link to a `world_location_news` document. The
preferred URL to render on the website is constructed from the base_path
of a world_location_news document, otherwise a URL is guessed from the
title of a world_location document.

Co-authored-by: Duncan Garmonsway <duncan.garmonsway@digital.cabinet-office.gov.u>