Add a content_store param to GET /v2/content/:id
There's a get_live_content method in gds-api adapters which would be
much simpler if this parameter were supported.

This would potentially also let us remove the version parameter, which
is only used by the get_live_content method and collections-publisher -
in both cases as a way to get live content (rather than a specific
old version of content).