Remove the version param on GET /v2/content/:id
Now that we've got the content_store parameter and we're using that
instead of version in the get_live_content method and in collections
publisher, we can remove the version parameter from the API endpoint.

To make absolutely sure that we haven't forgotten a call site and ended
up with subtly broken behaviour, I'm making providing a version
parameter an error (rather than silently ignoring the parameter).

The eventual aim of this is to remove the superseded state from the
query, so that we can optimise the way we store superseded editions
without having to worry about this use case. To reduce the number of
changes in each deployment, this commit / PR is just removing version
and we'll get rid of superseded states from the query afterwards.