Introduce /graphql/draft draft content endpoint
Which behaves a lot like the /graphql/content live content endpoint,
only with `with_drafts: true`.

I've left all the duplication here because it's tricky to see where to
begin with that. If we settle on having 2 separate controller actions
like this, then we can spend the time to make it easier and less error
prone to work with.

One consequence of the duplication is that it includes another file
access Brakeman warning to ignore (as the code that we duplicated did).

I've introduced a spec example here that demonstrates how this
endpoint's use of with_drafts=true enables a live edition to link to
a draft edition and vice versa. I've put it in its own separate
integration spec because it seems less focussed on endpoint behaviour
than the controller specs since it covers QueryType and dataloader
behaviour.

Co-authored-by: Ynda Jas <yndajas@gmail.com>