Replace QueryType content_store with with_drafts
As with the dataloader sources, this content_store setting was never
used and it defaulted to "live".

A query with `content_store: "live"` always returned a live edition or
no edition at all. That's the same behaviour as `with_drafts: false`.

A query with `content_store: "draft"` always returned a draft edition or
no edition. `with_drafts: true` doesn't behave like that because it can
return a draft edition or a live edition or no edition, with that order
of precedence. In Content Store land, there's a draft content item (i.e.
in Draft Content Store) for every published content item (i.e. in Live
Content Store). This `with_drafts` fallback behaviour is intended to
mimick the user-visible results of that.

Unlike the dataloader sources, which have to consider a linked-to
edition's state, QueryType looks up the root edition by base_path, which
will only have 0 or 1 exact matches for each of `content_store: "live"`
and `content_store: "draft"`.