Replace QueryType content_store with with_drafts
As with the dataloader sources, this content_store setting was never
used and it defaulted to "live".

A query with `content_store: "live"` always returned a live edition or
no edition at all. That's the same behaviour as `with_drafts: false`.

A query with `content_store: "draft"` always returned a draft edition or
no edition. `with_drafts: true` doesn't behave like that exactly, it can
return a draft edition or a live edition or no edition, with that order
of precedence.