WIP default fields order, no trailing newline
The JS version of the add-default-fields script has non-alphbetically
ordered default field names and removed the trailing newlines from the
query files.

I'm just adding these changes to the Ruby version of the script to make
sure that its ourput is identical to that of the JS version.

They do seem like minor improvements, though, so they could be
backported into the JS