Bump draft factory's default user_facing_version
The `live_edition` and `draft_edition` (alias `edition`) spec factories
were both defaulting to a user_facing_version of 1.

Since a document's live and draft editions are required to have
different user_facing_versions, this new default of 2 enables a spec
example to create both states of editions for a single document without
introducing the noise of explicitly setting this attribute for one or
both of them.

We can rethink this if it begins to make other tests more awkward in
future.

Note: that the unpublished_edition factory needed updating here too.
This was to prevent its default user_facing_version from changing since
it inherits from the factory we've changed.